Unit tests - Contact
There are two separated files because for one test we have to mock the useReducer hook and for others we don't want to. In order to have all theses tests into one file, we need to use the spyOn() jest function to restore orignal implementation of the hook. However, at this point of time, we failed to use it correctly in this specific case. If we succeed to transform the test in Contact2.test.js using spyOn() instead of jest.fn(), then it will be OK.